refactored admin component
