+++ src/client/app/containers/admin/PreferencesContainer.js
@@ -3,9 +3,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { connect } from 'react-redux';
-import _ from 'lodash';
-import AdminComponent from '../components/AdminComponent';
-import { updateDisplayTitle, updateDefaultChartToRender, toggleDefaultBarStacking, submitPreferencesIfNeeded, updateSelectedMeter } from '../actions/admin';
+import PreferencesComponent from '../../components/admin/PreferencesComponent';
+import { updateDisplayTitle, updateDefaultChartToRender, toggleDefaultBarStacking, submitPreferencesIfNeeded} from '../../actions/admin';
 
 function mapStateToProps(state) {
 	return {
@@ -13,14 +12,11 @@ function mapStateToProps(state) {
 		defaultChartToRender: state.admin.defaultChartToRender,
 		defaultBarStacking: state.admin.defaultBarStacking,
 		disableSubmitPreferences: state.admin.submitted,
-		meters: _.sortBy(_.values(state.meters.byMeterID).map(meter => ({ value: meter.id, label: meter.name.trim() })), 'name'),
-		selectedImportMeter: state.admin.selectedMeter
 	};
 }
 
 function mapDispatchToProps(dispatch) {
 	return {
-		updateSelectedImportMeter: meterID => dispatch(updateSelectedMeter(meterID)),
 		updateDisplayTitle: displayTitle => dispatch(updateDisplayTitle(displayTitle)),
 		updateDefaultGraphType: defaultChartToRender => dispatch(updateDefaultChartToRender(defaultChartToRender)),
 		toggleDefaultBarStacking: () => dispatch(toggleDefaultBarStacking()),
@@ -28,4 +24,5 @@ function mapDispatchToProps(dispatch) {
 	};
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(AdminComponent);
+export default connect(mapStateToProps, mapDispatchToProps)(PreferencesComponent);
+
